% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qle-package.R
\docType{data}
\name{matclust}
\alias{matclust}
\title{Matern cluster process data}
\format{A list object named `\code{matclust}` which consists of
  \itemize{
	   \item{qsd}{ initial quasi-likelihood approximation model}
    \item{OPT}{ the results of estimation by \code{\link{qle}}}
    \item{Stest}{ score test results }
  }}
\usage{
data(matclust)
}
\description{
A data set of quasi-likelihood estimation results of estimating the parameters of a Matern cluster
point process model. In the vignette we apply our method to the `\code{redwood}` data set from the
package \code{spatstat}.
}
\author{
M. Baaske
}
\keyword{datasets}
