% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qle-package.R
\docType{data}
\name{mm1q}
\alias{mm1q}
\title{QLE estimation results of M/M/1 queue}
\format{A list named `\code{mm1q}` with elements 
\itemize{
	   \item{qsd}{ initial quasi-likelihood approximation model}
    \item{OPT}{ the results of estimation by \code{\link{qle}}}
    \item{Stest}{ score test results }
	   \item{OPTS}{ results from simulation study, see the vignette}
	   \item{Stest0}{ Score test after estimating the model parameter }
	   \item{tet0}{ original parameter value}
	   \item{obs0}{ generated observed statistics for simulation study}
  }}
\usage{
data(mm1q)
}
\description{
The results of estimating the parameter of M/M/1 queue by Quasi-likelihood.
}
\author{
M. Baaske
}
\keyword{datasets}
