% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qle-package.R
\docType{data}
\name{OPT}
\alias{OPT}
\title{QLE estimation results of the normal model}
\format{A list named `\code{OPT}` of class \code{\link{qle}}, see function \code{\link{qle}}}
\usage{
data(qleresult)
}
\description{
The results of estimating the parameters of the normal model by Quasi-likelihood.
}
\author{
M. Baaske
}
\keyword{datasets}
