% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsOpt.R
\name{print.qle}
\alias{print.qle}
\title{print results of class \code{qle}}
\usage{
\method{print}{qle}(x, pl = 2, digits = 5, ...)
}
\arguments{
\item{x}{object of class \code{qle} from \code{\link{qle}}}

\item{pl}{numeric (positive) value, the print level (higher values give more information)}

\item{digits}{number of digits to display}

\item{...}{ignored, additional arguments}
}
\description{
S3 method to print the results from \code{\link{qle}}.
}
