% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qleTest.R
\name{print.qleTest}
\alias{print.qleTest}
\title{print \code{qleTest} results}
\usage{
\method{print}{qleTest}(x, pl = 1, digits = 5, ...)
}
\arguments{
\item{x}{object of class \code{qleTest} from \code{\link{qleTest}}}

\item{pl}{ignored}

\item{digits}{number of (significant) digits}

\item{...}{ignored}
}
\description{
print the results from \code{\link{qleTest}}
}
