% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qle-package.R
\docType{data}
\name{qsd}
\alias{qsd}
\title{A normal model}
\format{A list object named `\code{qsd}` of class \code{\link{QLmodel}} with additional elements
   \itemize{
	   \item{simfn}{ simulation function }
    \item{sim}{ simulation results at design points, class `\code{simQL}`}
	  }}
\usage{
data(normal)
}
\description{
A statistical model of random numbers
}
\details{
This is a pedagogic example of a simulated data set for quasi-likelihood estimation using
normally distributed random numbers. The model outcome is a vector of summary statistics, that is,
simply the median and mean average deviation of \code{n=10} random numbers, which is evaluated at the
model parameter \eqn{\theta=(\mu,\sigma)} with mean \eqn{\mu} and standard deviation \eqn{\sigma} as
the parameters of the normal distribution. We estimate the model parameter given a specific
"observation" of those summary statistics. Clearly, maximum likelihood estimation would be the 
method of first choice if we had a real sample of observations. However, this example is used to demonstrate
the basic workflow of estimating the model parameter. We use this model as a standard example in the package
documentation.
}
\author{
M. Baaske
}
\keyword{datasets}
